Moved dialogs to a different folder
