+++ Antrashare_ui/src/app/components/news-feed/story/comment-list/comment-dialog.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { CommentListComponent } from './comment-list.component';
+import { CommentListComponent } from './comment-dialog.component';
 
 describe('CommentListComponent', () => {
   let component: CommentListComponent;
@@ -8,9 +8,9 @@ describe('CommentListComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ CommentListComponent ]
+      declarations: [CommentListComponent]
     })
-    .compileComponents();
+      .compileComponents();
   });
 
   beforeEach(() => {
