+++ Antrashare_ui/src/app/components/news-feed/story/story-card/story-card.component.ts
@@ -2,7 +2,7 @@ import { Component, Input, OnInit } from '@angular/core';
 import { Subscription } from 'rxjs';
 import { MatDialog } from '@angular/material/dialog';
 import { newsStory } from 'src/app/models/newsStory.models';
-import { CommentListComponent } from '../comment-list/comment-list.component';
+import { CommentListComponent } from '../comment-list/comment-dialog.component';
 import { newsFeedService } from 'src/app/services/newsfeed/newsfeed.service';
 import { AuthService } from 'src/app/auth/auth.service';
