+++ Antrashare_ui/src/app/components/settings/settings.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { MatDialog, MatDialogRef } from '@angular/material/dialog';
-import { LogOutDialogComponent } from './logout-dialog/logout-dialog.component';
+import { LogOutDialogComponent } from '../../dialogs/logout-dialog/logout-dialog.component';
 @Component({
   selector: 'app-settings',
   templateUrl: './settings.component.html',
