+++ Antrashare_ui/src/app/dialogs/timeout-dialog/timeout-dialog.component.ts
